The step-by-step transition to the admin panel image we will use has begun (login and index). Firebase connections are handled in 3 stages. These stages are respectively; 1- realtime database 2- cloud firestore 3- firebase auth
File storage and admin panel parts are left to wait for us next. When we progress through these stages, we can now add the category and product titles needed for the menu as stock, and we can finish adding the users who will edit these sections on a role-based basis.